fixed a padding issue and allowed the warning again

+++ lib/include/ecs/config.hpp
@@ -3,7 +3,7 @@
 
 namespace ecs {
 
-typedef unsigned entity_id;
+typedef unsigned long entity_id;
 
 }
